requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0

# CLIP / Image processing
Pillow>=10.0.0

# Data
numpy>=1.24.0
pandas>=2.0.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# UI
gradio>=4.20.0

# Cloud storage
google-cloud-storage>=2.14.0

# Qwen2-VL dependencies
qwen-vl-utils>=0.0.2
